[IT-3021] Add 'priority-codes' parameter
Add a 'priority-codes' query-string parameter to move certain codes from
the chart of accounts to the top of the output.

Replace the 'enable_code_filter' parameter with debug logging output
and a refactor to ensure that the code processing always occurs. This
also renames 'filter_chart' to 'limit_chart' for clarity as it now only
processes the 'limit' parameter.

This will allow us to choose which tags will be listed in the Service
Catalog drop-down menu.