[IT-3072] Add requests timeout
This lambda typically runs in 2-3 seconds, but has recently started
hitting the 15-second AWS Lambda timeout.

Add a timeout when making requests to the upstream API to prevent the
AWS Lambda timeout from killing the process before we can log out.

Also add a timeout to the logout command in case that is hanging, to
allow the lambda to complete the current run and potentially cache
the result in cloudfront.