[IT-3316] Change upstream API routes
The current chart of accounts API route is returning an empty list,
replace the failing API route with two new ones that can successfully
return valid data.

In order to filter on the "Program" segment in this new API route,
we must first look up the segment ID based on its name, creating
a need for an additional API request.