[IT-4485] Refactor for arbitrary chart segments
Prepare for a new API endpoint that will require a different segment
of the chart of accounts.

The pre-existing endpoints used the Program segment of the chart of
accounts, but the new endpoint uses the GL segment of the chart of
accounts.

Refactor to allow arbitrary segments of the chart of accounts.